build: streamline examples build scripts
