+++ examples/package.json
@@ -5,10 +5,10 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "build": "npm run -s build:metadata && cross-env NODE_ENV=production node build.mjs",
-    "build:metadata": "node ./scripts/build-metadata.mjs",
-    "build:thumbnails": "node ./scripts/build-thumbnails.mjs",
-    "clean": "node ./scripts/clean.mjs",
+    "build": "cross-env NODE_ENV=production node build.mjs",
+    "build:metadata": "node build.mjs --metadata",
+    "build:thumbnails": "node build.mjs --thumbnails",
+    "clean": "node build.mjs --clean",
     "develop": "npm run -s build:metadata && cross-env NODE_ENV=development vite",
     "lint": "eslint .",
     "serve": "vite preview",
