--- examples/scripts/clean.mjs
@@ -1,4 +1,0 @@
-import fs from 'fs';
-
-fs.rmSync('dist', { recursive: true, force: true });
-fs.rmSync('cache', { recursive: true, force: true });
