+++ examples/utils/metadata.mjs
@@ -1,23 +1,20 @@
-import fs from 'fs';
-import path from 'path';
+import fs from 'node:fs';
+import path from 'node:path';
 
+import { parseConfig } from './example-source.mjs';
+import { bold, createdLog, startLog, writeLog, YELLOW } from './log.mjs';
 import { toKebabCase } from '../src/app/strings.mjs';
-import { bold, createdLog, startLog, writeLog, YELLOW } from '../utils/log.mjs';
-import { parseConfig } from '../utils/utils.mjs';
 
 const ROOT_PATH = 'src/examples';
 const OUTPUT = 'cache/metadata.mjs';
 
 /**
- * @type {{
- *      path: string,
- *      categoryKebab: string,
- *      exampleNameKebab: string,
- *      hidden: boolean
- * }[]}
+ * @typedef {object} ExampleMetadata
+ * @property {string} path - The example directory.
+ * @property {string} categoryKebab - The category name.
+ * @property {string} exampleNameKebab - The example name.
+ * @property {boolean} hidden - True if hidden from the production sidebar.
  */
-const exampleMetaData = [];
-const hiddenExamples = [];
 
 /**
  * @param {object} obj - The object.
@@ -28,25 +25,30 @@ const objStringify = (obj) => {
 };
 
 /**
- * @param {string} path - The directory path.
+ * @param {string} dir - The directory path.
  * @returns {string[]} - The file names in the directory.
  */
-const getDirFiles = (path) => {
-    if (!fs.existsSync(path)) {
+const getDirFiles = (dir) => {
+    if (!fs.existsSync(dir)) {
         return [];
     }
-    const stats = fs.statSync(path);
+    const stats = fs.statSync(dir);
     if (!stats.isDirectory()) {
         return [];
     }
-    return fs.readdirSync(path);
+    return fs.readdirSync(dir);
 };
 
-
-const main = () => {
+/**
+ * @returns {void} no return value.
+ */
+export const buildMetadata = () => {
     startLog('metadata', OUTPUT);
     const start = performance.now();
     const categories = getDirFiles(ROOT_PATH);
+    /** @type {ExampleMetadata[]} */
+    const exampleMetaData = [];
+    const hiddenExamples = [];
 
     categories.forEach((category) => {
         const categoryPath = path.resolve(`${ROOT_PATH}/${category}`);
@@ -60,9 +62,9 @@ const main = () => {
             const examplePath = path.resolve(`${categoryPath}/${exampleFile}`);
             const exampleName = exampleFile.split('.').shift() ?? '';
             const exampleNameKebab = toKebabCase(exampleName);
-
             const config = parseConfig(fs.readFileSync(examplePath, 'utf-8'));
             const hidden = !!config.HIDDEN;
+
             if (hidden) {
                 hiddenExamples.push(`${categoryKebab}/${exampleNameKebab}`);
             }
@@ -76,10 +78,7 @@ const main = () => {
         });
     });
 
-    if (!fs.existsSync('cache')) {
-        fs.mkdirSync('cache');
-    }
-
+    fs.mkdirSync('cache', { recursive: true });
     fs.writeFileSync(OUTPUT, `export const exampleMetaData = ${objStringify(exampleMetaData)};\n`);
 
     if (hiddenExamples.length) {
@@ -89,4 +88,3 @@ const main = () => {
 
     createdLog(OUTPUT, performance.now() - start);
 };
-main();
